minor refactoring
